+++ LibOS/shim/src/sys/shim_wrappers.c
@@ -15,7 +15,7 @@
 #include "shim_table.h"
 #include "shim_utils.h"
 
-ssize_t shim_do_readv(int fd, const struct iovec* vec, int vlen) {
+long shim_do_readv(int fd, const struct iovec* vec, int vlen) {
     if (!vec || test_user_memory((void*)vec, sizeof(*vec) * vlen, false))
         return -EINVAL;
 
@@ -77,7 +77,7 @@ out:
  * actually written. Otherwise, it shall return a value of -1, the file-pointer
  * shall remain unchanged, and errno shall be set to indicate an error
  */
-ssize_t shim_do_writev(int fd, const struct iovec* vec, int vlen) {
+long shim_do_writev(int fd, const struct iovec* vec, int vlen) {
     if (!vec || test_user_memory((void*)vec, sizeof(*vec) * vlen, false))
         return -EINVAL;
